requirements.txt
oci>=2.161.0
python-dotenv>=1.0.1
